Added battleships, destroyers, bombers, and fighters
It was a dark and stormy night, the rain fell in torrents - except at
occasional intervals, when it was checked by a violent gust of wind
which swept up the streets (for it is in London that our scene lies),
rattling along the housetops and fiercely agitating the scanty flame of
the lamps that struggled against the darkness. Through the miserable
torrent our hero Ali bravely creates scalable vector graphics for
battleships, bombers, fighters and destroyers. Some of the battleships
double as destroyers. Where did Ali put his coffee... bugger.