Added carriers, subs, transports
"Commit to master". Such brutal, uncultured, unrefined language. It's
this sort of willing submission to the upperclassmen that undermines the
roots of our endeavours to create a classless utopia. How can we expect
to free ourselves from the shackles of our bourgeoisie society if
elements of classist language continue permeate all aspects of life,
online and off. I WILL NOT BE A SLAVE.

Note that one of the transports is a ground transport while the other is
a landing craft.