Add amphibious travel - combat
Land units at sea, and sea units on land, are considered
to be noncombatants. They do not participate in the fight,
and will be destroyed if the enemy wins. Factories are
neutral noncombatants, and will be captured by the victor.

Removed static lists of conflicts - instead, list is generated
on the fly when required.